Refactor: move all icons to commons components
