+++ src/js/components/ManageMenu/CommunityCloud/SearchResult.tsx
@@ -8,7 +8,7 @@ import { Button } from '../../common/Button'
 import { CloudSnippetDownloadButton } from '../../common/cloud/CloudSnippetDownloadButton'
 import { CloudSnippetPreviewModal } from '../../common/cloud/CloudSnippetPreviewModal'
 import { CloudStatusIndicator } from '../../common/cloud/CloudStatusBadge'
-import { CloudUpdateIcon } from '../../common/icons/CloudIcons'
+import { CloudUpdateIcon } from '../../common/icons/CloudUpdateIcon'
 import { SnippetCard } from '../../common/SnippetCard'
 import { useCloudSearch } from './WithCloudSearchContext'
 import type { ReactNode } from 'react'
