+++ src/js/components/common/Toolbar.tsx
@@ -3,7 +3,11 @@ import classnames from 'classnames'
 import React, { useState } from 'react'
 import { isLicensed, shouldShowUpsell } from '../../utils/screen'
 import { buildUrl, fetchQueryParam } from '../../utils/urls'
-import { BlueprintIcon, CommunityIcon, LibraryIcon, SettingsIcon, SnippetsIcon } from './icons/ToolbarIcons'
+import { BlueprintIcon } from './icons/BlueprintIcon'
+import { CommunityIcon } from './icons/CommunityIcon'
+import { LibraryIcon } from './icons/LibraryIcon'
+import { SettingsIcon } from './icons/SettingsIcon'
+import { SnippetsIcon } from './icons/SnippetsIcon'
 import { UpsellDialog } from './UpsellDialog'
 import type { ReactNode } from 'react'
