--- src/js/components/common/icons/CloudIcons.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import type { SVGProps } from 'react'
-
-export const CloudUpdateIcon: React.FC<SVGProps<SVGSVGElement>> = props =>
-	<svg width="19" height="19" viewBox="0 0 19 19" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
-		<path d="M9.17636 13.7674H5.09053C3.12371 13.7643 1.5293 12.2296 1.5293 10.3362C1.5293 8.44351 3.12371 6.90874 5.09053 6.90874C5.39106 5.56133 6.46242 4.46168 7.90083 4.02351C9.33848 3.58609 10.926 3.87592 12.0639 4.78821C13.2018 5.69821 13.7172 7.08768 13.4174 8.4351H14.1745C15.2298 8.4351 16.1421 9.0568 16.568 9.95686M14.5293 12.2349V16.8232M12.2352 14.529L14.5293 16.8232L16.8234 14.529"
-			stroke="#FF9800"
-			strokeWidth="1.41177"
-			strokeLinecap="round"
-			strokeLinejoin="round"
-		/>
-	</svg>
