+++ src/js/components/common/icons/MinimiseIcon.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import type { SVGProps } from 'react'
 
-export const MinimiseIcon = (props: SVGProps<SVGSVGElement>) =>
+export const MinimiseIcon: React.FC<SVGProps<SVGSVGElement>> = props =>
 	<svg width="18" height="18" viewBox="0 0 18 18" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
 		<path d="M3.75 6.75H6.75V3.75" />
 		<path d="M2.25 2.25L6.75 6.75L2.25 2.25Z" />
