+++ src/js/components/common/icons/TableViewIcon.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import type { SVGProps } from 'react'
+
+export const TableViewIcon: React.FC<SVGProps<SVGSVGElement>> = props =>
+	<svg width="18" height="18" viewBox="0 0 18 18" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+		<path
+			d="M2.5 4.5H15.5M2.5 9H15.5M2.5 13.5H15.5"
+			stroke="currentColor"
+			strokeWidth="1.8"
+			strokeLinecap="round"
+		/>
+	</svg>
